replaced PsalmDeleteModal with an universal Modal

+++ src/components/home/ProjectList.vue
@@ -5,13 +5,13 @@
 		<div class="ml-2 mr-6 grid grid-cols-[1fr,2rem] font-semibold border-b border-gray-400">
 			<div class="px-1" title="Titel">Titel</div>
 		</div>
-		<div class="lg:overflow-y-scroll scrollbar-p-2 max-h-[calc(100vh-11.25rem)] mx-2 lg:mr-0">
+		<div class="lg:overflow-y-scroll scrollbar-p-2 max-h-[calc(100vh-11.3125rem)] mx-2 lg:mr-0">
 			<ul>
 				<li class="grid grid-cols-[1fr,2rem] grid-rows-[auto,auto] border-b border-gray-400 last:border-b-0" v-for="project in projectsArray" :key="project.id">
 					<!-- PROJECTS INPUTS -->
-					<router-link :to="`/project/${project.id}`" v-if="!editMode" class="my-1 px-1 hover:text-secondary min-h-[1.5rem] border border-transparent overflow-ellipsis overflow-hidden whitespace-nowrap" :title="project.title">{{
-						project.title
-					}}</router-link>
+					<router-link :to="`/project/${project.id}`" v-if="!editMode" class="my-1 px-1 hover:text-secondary min-h-[1.5rem] border border-transparent overflow-ellipsis overflow-hidden whitespace-nowrap" :title="project.title">
+						{{ project.title }}
+					</router-link>
 					<PsalmInput v-if="editMode" class="my-1 w-full" type="text" v-model.trim="project.title" placeholder="Titel" :title="project.title" />
 					<PsalmDeleteButton @click="triggerDeleteModal(project)" />
 				</li>
@@ -30,8 +30,8 @@
 			<PsalmButton title="Doodle Liste importieren" class="bg-primary" icon="import" @click="showDoodleImportModal = true" />
 		</div>
 
-		<!-- DELETE MODAL -->
-		<PsalmDeleteModal v-if="showDeleteModal" type="project" :object-to-delete="projectToDelete" @confirm="deleteProject" @cancel="showDeleteModal = false" />
+		<!-- MODAL -->
+		<PsalmModal v-if="modal.show && modal.type === 'DELETE_PROJECT'" @confirm="deleteProject" />
 
 		<!-- DOODLE IMPORT MODAL -->
 		<DoodleImportModal v-if="showDoodleImportModal" @import="importProject" @cancel="showDoodleImportModal = false" />
@@ -42,50 +42,54 @@
 	import store from "@/store";
 	import { Component, Vue } from "vue-property-decorator";
 	import DoodleImportModal from "@/components/home/DoodleImportModal.vue";
-	import PsalmDeleteModal from "@/components/common/PsalmDeleteModal.vue";
+	import PsalmModal from "@/components/common/PsalmModal.vue";
 	import PsalmButton from "@/components/common/PsalmButton.vue";
 	import PsalmDeleteButton from "@/components/common/PsalmDeleteButton.vue";
 	import PsalmCard from "@/components/common/PsalmCard.vue";
 	import PsalmIcon from "@/components/common/PsalmIcon.vue";
 	import PsalmInput from "@/components/common/PsalmInput.vue";
 	import { newProject, Project } from "@/models/interfaces/Project";
-	import { newID, pathExists } from "@/utils/utils";
+	import { equal, newID, pathExists } from "@/utils/utils";
 	import { removeFile, writeFile } from "@tauri-apps/api/fs";
+	import { Modal } from "@/models/interfaces/Modal";
+	import { ModalType } from "@/models/enums/ModalType";
 
 	@Component({
 		name: "ProjectList",
-		components: { DoodleImportModal, PsalmDeleteModal, PsalmButton, PsalmDeleteButton, PsalmCard, PsalmIcon, PsalmInput },
+		components: { DoodleImportModal, PsalmModal, PsalmButton, PsalmDeleteButton, PsalmCard, PsalmIcon, PsalmInput },
 	})
 	export default class ProjectList extends Vue {
 		tempProjects: Map<string, Project> = new Map();
 		projectsArray: Array<Project> = [];
 		editMode = false;
-		showDeleteModal = false;
-		projectToDelete = newProject(newID());
 		showDoodleImportModal = false;
 
+		get modal(): Modal {
+			return store.state.modal;
+		}
+
 		created(): void {
-			this.tempProjects = new Map(store.state.projects);
+			// this.tempProjects = new Map(store.state.projects);
+			this.tempProjects = new Map(JSON.parse(JSON.stringify(Array.from(store.state.projects))));
 			this.updateProjectsArray();
 		}
 
 		triggerDeleteModal(projectToDelete: Project): void {
-			this.projectToDelete = projectToDelete;
-			this.showDeleteModal = true;
+			store.commit("showModal", { type: ModalType.DELETE_PROJECT, content: projectToDelete });
 		}
 
 		async deleteProject(): Promise<void> {
-			this.tempProjects.delete(this.projectToDelete.id);
+			const projectToDelete = this.modal.content as Project;
 
-			if (await pathExists("data\\projects", `data\\projects\\${this.projectToDelete.id}.json`)) {
-				await removeFile(`data/projects/${this.projectToDelete.id}.json`);
+			this.tempProjects.delete(projectToDelete.id);
+
+			if (await pathExists("data\\projects", `data\\projects\\${projectToDelete.id}.json`)) {
+				await removeFile(`data/projects/${projectToDelete.id}.json`);
 			}
 
-			// store.commit("updateProjects", this.tempProjects);
-			// this.updateProjectsArray();
 			await this.saveProjects();
 			store.commit("showToast", "deleted");
-			this.showDeleteModal = false;
+			store.commit("resetModal");
 		}
 
 		addProject(): void {
@@ -99,21 +103,21 @@
 		}
 
 		async saveProjects(): Promise<void> {
-			// SEARCH FOR EDITED PROJECTS
-			// temp
+			this.tempProjects.forEach(async (value, key) => {
+				// SEARCH FOR EDITED PROJECTS
+				if (!store.state.projects.get(key) || !equal(value, store.state.projects.get(key))) {
+					// DELETE OLD JSON FILE
+					if (await pathExists("data\\projects", `data\\projects\\${key}.json`)) {
+						await removeFile(`data/projects/${key}.json`);
+					}
+
+					// WRITE NEW JSON FILE
+					await writeFile({ contents: JSON.stringify(value), path: `data/projects/${key}.json` });
+				}
+			});
 
 			// PUSH TEMP-PROJECTS TO STORE
 			store.commit("updateProjects", this.tempProjects);
-
-			this.tempProjects.forEach(async (value, key) => {
-				// DELETE JSON FILE
-				if (await pathExists("data\\projects", `data\\projects\\${key}.json`)) {
-					await removeFile(`data/projects/${key}.json`);
-				}
-
-				// WRITE JSON FILE
-				await writeFile({ contents: JSON.stringify(value), path: `data/projects/${key}.json` });
-			});
 
 			// CREATE ARRAY FOR v-for LOOP
 			this.updateProjectsArray();
